fix: prevent problèmes techniques

+++ packages/api/src/modules/events/resolvers/mutation.delete-event.ts
@@ -1,4 +1,7 @@
 import { builder, prisma } from '#lib';
+import { GraphQLError } from 'graphql';
+import { unlink } from 'node:fs/promises';
+import { join } from 'node:path';
 
 builder.mutationField('deleteEvent', (t) =>
   t.field({
@@ -16,6 +19,24 @@ builder.mutationField('deleteEvent', (t) =>
       );
     },
     async resolve(_, { id }, { user }) {
+      const event = await prisma.event.findUniqueOrThrow({
+        where: { id },
+        include: { beneficiary: true },
+      });
+
+      if (event.beneficiary)
+        {throw new GraphQLError(
+          'Tu ne peux pas supprimer un événement où des places sont réservées.',
+        );}
+
+      const { pictureFile } = await prisma.event.findUniqueOrThrow({
+        where: { id },
+        select: { pictureFile: true },
+      });
+
+      const root = new URL(process.env.STORAGE).pathname;
+
+      if (pictureFile) await unlink(join(root, pictureFile));
       await prisma.event.delete({
         where: { id },
       });
